internal/jem: refactor
Separate the purely database operations from the shared business logic.
