+++ cmd/jaas-model/modelcmd/remove_test.go
@@ -31,7 +31,7 @@ func (s *removeSuite) TestRemoveModel(c *gc.C) {
 	stdout, stderr, code = run(c, c.MkDir(), "list")
 	c.Assert(code, gc.Equals, 0, gc.Commentf("stderr: %s", stderr))
 	c.Assert(stderr, gc.Equals, "")
-	c.Assert(stdout, gc.Equals, "bob/foo\n")
+	c.Assert(stdout, gc.Equals, "")
 }
 
 func (s *removeSuite) TestRemoveController(c *gc.C) {
@@ -72,7 +72,7 @@ func (s *removeSuite) TestRemoveController(c *gc.C) {
 	stdout, stderr, code = run(c, c.MkDir(), "list")
 	c.Assert(code, gc.Equals, 0, gc.Commentf("stderr: %s", stderr))
 	c.Assert(stderr, gc.Equals, "")
-	c.Assert(stdout, gc.Equals, "bob/bar\n")
+	c.Assert(stdout, gc.Equals, "")
 }
 
 func (s *removeSuite) TestRemoveMultipleModels(c *gc.C) {
@@ -89,12 +89,12 @@ func (s *removeSuite) TestRemoveMultipleModels(c *gc.C) {
 	stdout, stderr, code = run(c, c.MkDir(), "remove", "bob/foo", "bob/foo-1")
 	c.Assert(code, gc.Equals, 1, gc.Commentf("stderr: %s", stderr))
 	c.Assert(stdout, gc.Equals, "")
-	c.Assert(stderr, gc.Matches, `cannot remove bob/foo: DELETE http://.*/v2/model/bob/foo: cannot remove model "bob/foo" because it is a controller`+"\n")
+	c.Assert(stderr, gc.Matches, `cannot remove bob/foo: DELETE http://.*/v2/model/bob/foo: model "bob/foo" not found`+"\n")
 
 	stdout, stderr, code = run(c, c.MkDir(), "list")
 	c.Assert(code, gc.Equals, 0, gc.Commentf("stderr: %s", stderr))
 	c.Assert(stderr, gc.Equals, "")
-	c.Assert(stdout, gc.Equals, "bob/foo\n")
+	c.Assert(stdout, gc.Equals, "")
 }
 
 func (s *removeSuite) TestRemoveVerbose(c *gc.C) {
@@ -116,7 +116,7 @@ func (s *removeSuite) TestRemoveVerbose(c *gc.C) {
 	stdout, stderr, code = run(c, c.MkDir(), "list")
 	c.Assert(code, gc.Equals, 0, gc.Commentf("stderr: %s", stderr))
 	c.Assert(stderr, gc.Equals, "")
-	c.Assert(stdout, gc.Equals, "bob/foo\n")
+	c.Assert(stdout, gc.Equals, "")
 }
 
 var removeErrorTests = []struct {
