+++ cmd/jaas-model/modelcmd/revoke_test.go
@@ -18,12 +18,14 @@ var _ = gc.Suite(&revokeSuite{})
 func (s *revokeSuite) TestRevoke(c *gc.C) {
 	s.idmSrv.SetDefaultUser("bob")
 
-	// First add a controller. This also adds an model that we can
-	// alter for our test.
+	// First add a controller.
 	stdout, stderr, code := run(c, c.MkDir(), "add-controller", "bob/foo")
 	c.Assert(code, gc.Equals, 0, gc.Commentf("stderr: %s", stderr))
 	c.Assert(stdout, gc.Equals, "")
 	c.Assert(stderr, gc.Equals, "")
+
+	// Create a model to use with our test.
+	s.addEnv(c, "bob/foo", "bob/foo", "cred1")
 
 	// Check that alice can't get controller or model.
 	aliceClient := s.jemClient("alice")
