--- internal/jem/juju.go
@@ -1,206 +1,0 @@
-// Copyright 2016 Canonical Ltd.
-
-package jem
-
-import (
-	"fmt"
-
-	cloudapi "github.com/juju/juju/api/cloud"
-	"github.com/juju/juju/api/modelmanager"
-	jujuparams "github.com/juju/juju/apiserver/params"
-	jujucloud "github.com/juju/juju/cloud"
-	"gopkg.in/errgo.v1"
-	"gopkg.in/juju/names.v2"
-	"gopkg.in/mgo.v2/bson"
-
-	"github.com/CanonicalLtd/jem/internal/apiconn"
-	"github.com/CanonicalLtd/jem/internal/mongodoc"
-	"github.com/CanonicalLtd/jem/params"
-)
-
-// CreateModelParams specifies the parameters needed to create a new
-// model using CreateModel.
-type CreateModelParams struct {
-	// Path contains the path of the new model.
-	Path params.EntityPath
-
-	// ControllerPath contains the path of the owning
-	// controller.
-	ControllerPath params.EntityPath
-
-	// Credential contains the name of the credential to use to
-	// create the model.
-	Credential params.Name
-
-	// Cloud contains the name of the cloud in which the
-	// model will be created.
-	Cloud params.Cloud
-
-	// Region contains the name of the region in which the model will
-	// be created. This may be empty if the cloud does not support
-	// regions.
-	Region string
-
-	// Attributes contains the attributes to assign to the new model.
-	Attributes map[string]interface{}
-}
-
-// CreateModel creates a new model as specified by p using conn.
-func (j *JEM) CreateModel(conn *apiconn.Conn, p CreateModelParams) (*mongodoc.Model, *jujuparams.ModelInfo, error) {
-	cred, err := j.Credential(p.Path.User, p.Cloud, p.Credential)
-	if err != nil {
-		return nil, nil, errgo.Mask(err, errgo.Is(params.ErrNotFound))
-	}
-	if err := j.updateControllerCredential(conn, cred); err != nil {
-		return nil, nil, errgo.Mask(err)
-	}
-	if err := j.credentialAddController(p.Path.User, p.Cloud, p.Credential, p.ControllerPath); err != nil {
-		return nil, nil, errgo.Mask(err)
-	}
-	// Create the model record in the database before actually
-	// creating the model on the controller. It will have an invalid
-	// UUID because it doesn't exist but that's better than creating
-	// an model that we can't add locally because the name
-	// already exists.
-	modelDoc := &mongodoc.Model{
-		Path:       p.Path,
-		Controller: p.ControllerPath,
-	}
-	if err := j.AddModel(modelDoc); err != nil {
-		return nil, nil, errgo.Mask(err, errgo.Is(params.ErrAlreadyExists))
-	}
-	mmClient := modelmanager.NewClient(conn.Connection)
-	m, err := mmClient.CreateModel(
-		string(p.Path.Name),
-		UserTag(p.Path.User).Id(),
-		string(p.Cloud),
-		p.Region,
-		CloudCredentialTag(p.Cloud, p.Path.User, p.Credential),
-		p.Attributes,
-	)
-	if err != nil {
-		// Remove the model that was created, because it's no longer valid.
-		if err := j.DB.Models().RemoveId(modelDoc.Id); err != nil {
-			logger.Errorf("cannot remove model from database after model creation error: %v", err)
-		}
-		return nil, nil, errgo.Notef(err, "cannot create model")
-	}
-	if err := mmClient.GrantModel(conn.Info.Tag.(names.UserTag).Id(), "admin", m.UUID); err != nil {
-		// TODO (mhilton) destroy the model?
-		return nil, nil, errgo.Notef(err, "cannot grant admin access")
-	}
-	// Now set the UUID to that of the actually created model.
-	if err := j.DB.Models().UpdateId(modelDoc.Id, bson.D{{"$set", bson.D{{"uuid", m.UUID}}}}); err != nil {
-		// TODO (mhilton) destroy the model?
-		return nil, nil, errgo.Notef(err, "cannot update model UUID in database, leaked model %s", m.UUID)
-	}
-	modelDoc.UUID = m.UUID
-	return modelDoc, &m, nil
-}
-
-// UpdateAndDeployCredential updates the specified credential in the
-// local database and then updates it on all controllers to which it is
-// deployed.
-func (j *JEM) UpdateCredential(cred *mongodoc.Credential) error {
-	if err := j.updateCredential(cred); err != nil {
-		return errgo.Notef(err, "cannot update local database")
-	}
-	c, err := j.Credential(cred.User, cred.Cloud, cred.Name)
-	if err != nil {
-		return errgo.Mask(err)
-	}
-	// TODO(mhilton) consider how to handle and recover from errors
-	// updating credentials in controllers better.
-	var firstError error
-	for _, ctl := range c.Controllers {
-		conn, err := j.OpenAPI(ctl)
-		if err != nil {
-			logger.Errorf("cannot open controller connection to %s: %s", ctl, err)
-			if firstError != nil {
-				firstError = err
-			}
-			continue
-		}
-		if err := j.updateControllerCredential(conn, cred); err != nil {
-			logger.Errorf("cannot update credential %s on %s: %s", cred.Id, ctl, err)
-			if firstError != nil {
-				firstError = err
-			}
-		}
-		conn.Close()
-	}
-	return errgo.Mask(firstError)
-}
-
-// updateControllerCredential uploads the given credential to conn.
-func (j *JEM) updateControllerCredential(conn *apiconn.Conn, cred *mongodoc.Credential) error {
-	cloudCredentialTag := CloudCredentialTag(cred.Cloud, cred.User, cred.Name)
-	cloudClient := cloudapi.NewClient(conn)
-	err := cloudClient.UpdateCredential(
-		cloudCredentialTag,
-		jujucloud.NewCredential(jujucloud.AuthType(cred.Type), cred.Attributes),
-	)
-	if err != nil {
-		return errgo.Notef(err, "cannot upload credentials")
-	}
-	return nil
-}
-
-// GrantModel grants the given access for the given user on the given model and updates the JEM database.
-func (j *JEM) GrantModel(conn *apiconn.Conn, model *mongodoc.Model, user params.User, access string) error {
-	client := modelmanager.NewClient(conn)
-	if err := client.GrantModel(UserTag(user).Id(), access, model.UUID); err != nil {
-		return errgo.Mask(err)
-	}
-	if err := j.Grant(j.DB.Models(), model.Path, user); err != nil {
-		// TODO (mhilton) What should be done with the changes already made to the controller.
-		return errgo.Mask(err)
-	}
-	return nil
-}
-
-// RevokeModel revokes the given access for the given user on the given model and updates the JEM database.
-func (j *JEM) RevokeModel(conn *apiconn.Conn, model *mongodoc.Model, user params.User, access string) error {
-	if err := j.Revoke(j.DB.Models(), model.Path, user); err != nil {
-		return errgo.Mask(err)
-	}
-	client := modelmanager.NewClient(conn)
-	if err := client.RevokeModel(UserTag(user).Id(), access, model.UUID); err != nil {
-		// TODO (mhilton) What should be done with the changes already made to JEM.
-		return errgo.Mask(err)
-	}
-	return nil
-}
-
-// DestroyModel destroys the specified model and removes it from the
-// database.
-//
-// Note that if the model is destroyed in its controller but
-// j.DeleteModel fails, a subsequent DestroyModel can can still succeed
-// because client.DestroyModel will succeed when the model doesn't exist.
-func (j *JEM) DestroyModel(conn *apiconn.Conn, model *mongodoc.Model) error {
-	client := modelmanager.NewClient(conn)
-	if err := client.DestroyModel(names.NewModelTag(model.UUID)); err != nil {
-		return errgo.Mask(err)
-	}
-	if err := j.DeleteModel(model.Path); err != nil {
-		return errgo.Mask(err)
-	}
-	return nil
-}
-
-// UserTag creates a juju user tag from a params.User
-func UserTag(u params.User) names.UserTag {
-	return names.NewUserTag(string(u) + "@external")
-}
-
-// CloudTag creates a juju cloud tag from a params.Cloud
-func CloudTag(c params.Cloud) names.CloudTag {
-	return names.NewCloudTag(string(c))
-}
-
-// CloudCredentialTag creates a juju cloud credential tag from the given
-// cloud, user and name.
-func CloudCredentialTag(cloud params.Cloud, user params.User, name params.Name) names.CloudCredentialTag {
-	return names.NewCloudCredentialTag(fmt.Sprintf("%s/%s@external/%s", cloud, user, name))
-}
