--- internal/v2/export_test.go
@@ -1,5 +1,0 @@
-package v2
-
-var (
-	SchemaForProviderType = &schemaForProviderType
-)
