+++ params/params.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/juju/httprequest"
 	"gopkg.in/errgo.v1"
-	"gopkg.in/juju/environschema.v1"
 )
 
 // SetControllerPerm holds the parameters for setting the ACL on a
@@ -238,16 +237,9 @@ type ControllerResponse struct {
 	// Path holds the path of the controller.
 	Path EntityPath `json:"path"`
 
-	// TODO perhaps these two fields should be put
-	// into the same type as returned from GetSchema.
-
 	// ProviderType holds the kind of provider used
 	// by the Controller.
 	ProviderType string `json:"provider-type,omitempty"`
-
-	// Schema holds the fields required to start
-	// a new model using the Controller.
-	Schema environschema.Fields `json:"schema,omitempty"`
 
 	// Location holds location attributes associated with the controller.
 	Location map[string]string `json:"location,omitempty"`
@@ -266,27 +258,6 @@ type ControllerResponse struct {
 type GetController struct {
 	httprequest.Route `httprequest:"GET /v2/controller/:User/:Name"`
 	EntityPath
-}
-
-// GetSchema holds parameters for getting a schema.
-type GetSchema struct {
-	httprequest.Route `httprequest:"GET /v2/schema"`
-
-	// Location constrains the controllers that will be used to
-	// fetch the schema.
-	//
-	// Note that the values in this should be passed in the
-	// URL query parameters.
-	Location map[string]string
-}
-
-// SchemaResponse holds the information returned by a GetSchema request.
-type SchemaResponse struct {
-	// ProviderType holds the kind of provider associated with the schema.
-	ProviderType string `json:"provider-type,omitempty"`
-
-	// Schema holds the fields required to start a new model.
-	Schema environschema.Fields `json:"schema,omitempty"`
 }
 
 // NewModelInfo holds the JSON body parameters
